Remove double warning in CLI  when config value is deprecated (#40319)
* Remove double warning in CLI  when config value is deprecated

There's a double warning when getting a value for a deprecated configuration. This happens because we first check if the config exists, which issues a warning and then getting the config also issues the same warning. Now that we don't error for non-existing configs, we should not check existence before getting a config value

* mark test_plugins_command as db_test

* mark test_cli_parser.py as db_test